@@ -13,45 +13,50 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.0.0",
-    "@angular/common": "^4.0.0",
-    "@angular/compiler": "^4.0.0",
-    "@angular/core": "^4.0.0",
-    "@angular/forms": "^4.0.0",
-    "@angular/http": "^4.0.0",
-    "@angular/platform-browser": "^4.0.0",
-    "@angular/platform-browser-dynamic": "^4.0.0",
-    "@angular/router": "^4.0.0",
-    "angular2-jwt": "^0.2.3",
+    "@angular/animations": "7.0.1",
+    "@angular/common": "7.0.1",
+    "@angular/compiler": "7.0.1",
+    "@angular/core": "7.0.1",
+    "@angular/forms": "7.0.1",
+    "@angular/http": "7.0.1",
+    "@angular/platform-browser": "7.0.1",
+    "@angular/platform-browser-dynamic": "7.0.1",
+    "@angular/router": "7.0.1",
+    "@auth0/angular-jwt": "^2.0.0",
+    "chart.js": "^2.7.3",
     "core-js": "^2.4.1",
     "enhanced-resolve": "^3.3.0",
     "font-awesome": "^4.7.0",
     "moment": "^2.22.2",
     "ng2-currency-mask": "^5.3.1",
     "ng2-toasty": "^4.0.3",
-    "primeng": "^4.1.0-rc.2",
-    "rxjs": "^5.1.0",
-    "zone.js": "^0.8.4",
-    "@angular/cli": "1.2.7",
-    "@angular/compiler-cli": "^4.0.0",
-    "@angular/language-service": "^4.0.0",
-    "@types/jasmine": "2.5.45",
+    "primeicons": "^1.0.0",
+    "primeng": "^6.1.5",
+    "rxjs": "^6.3.3",
+    "rxjs-compat": "^6.3.3",
+    "zone.js": "^0.8.26",
+    "@angular-devkit/build-angular": "~0.10.0",
+    "@angular/cli": "7.0.3",
+    "@angular/compiler-cli": "7.0.1",
+    "@angular/language-service": "7.0.1",
+    "@types/jasmine": "~2.8.3",
+    "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
-    "codelyzer": "~3.0.1",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
-    "karma-cli": "~1.0.1",
+    "codelyzer": "^4.0.1",
+    "jasmine-core": "~2.8.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~2.0.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "^1.2.1",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
     "protractor": "~5.1.2",
-    "ts-node": "~3.0.4",
-    "tslint": "~5.3.2",
-    "typescript": "~2.3.3"
+    "ts-node": "~4.1.0",
+    "tslint": "~5.9.1",
+    "typescript": "3.1.3"
   },
   "devDependencies": {
+
   },
   "engines": {
     "node": "10.10.0",
