@@ -1,11 +1,38 @@
+export class Estado {
+  codigo: number;
+  nome: string;
+}
+
+export class Cidade {
+  codigo: number;
+  nome: string;
+  estado = new Estado();
+}
+
 export class Endereco {
   logradouro: string;
   numero: string;
   complemento: string;
   bairro: string;
   cep: string;
-  cidade: string;
-  estado: string;
+  cidade = new Cidade();
+}
+
+export class Contato {
+  codigo: number;
+  nome: string;
+  email: string;
+  telefone: string;
+
+  constructor(codigo?: number,
+    nome?: string,
+    email?: string,
+    telefone?: string){
+      this.codigo = codigo;
+      this.nome = nome;
+      this.email = email;
+      this.telefone = telefone;
+    }
 }
 
 export class Pessoa {
@@ -14,6 +41,7 @@ export class Pessoa {
   endereco = new Endereco();
   ativo: true;
   empresa = new Empresa();
+  contatos = Array<Contato>();
 }
 
 export class Categoria {
