@@ -1,12 +1,13 @@
-import { SharedModule } from './../shared/shared.module';
-import { TooltipModule } from 'primeng/components/tooltip/tooltip';
-import { DataTableModule } from 'primeng/components/datatable/datatable';
-import { ButtonModule } from 'primeng/components/button/button';
-import { InputTextModule } from 'primeng/components/inputtext/inputtext';
 import { FormsModule } from '@angular/forms';
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 
+import { TooltipModule } from 'primeng/tooltip';
+import { TableModule } from 'primeng/table';
+import { ButtonModule } from 'primeng/button';
+import { InputTextModule } from 'primeng/inputtext';
+
+import { SharedModule } from './../shared/shared.module';
 import { EmpresasRoutingModule } from './empresas-routing.module';
 import { EmpresasPesquisaComponent } from './empresas-pesquisa/empresas-pesquisa.component';
 import { EmpresaCadastroComponent } from './empresa-cadastro/empresa-cadastro.component';
@@ -18,7 +19,7 @@ import { EmpresaCadastroComponent } from './empresa-cadastro/empresa-cadastro.co
 
     InputTextModule,
     ButtonModule,
-    DataTableModule,
+    TableModule,
     TooltipModule,
 
     SharedModule,
