@@ -1,6 +1,7 @@
 import { environment } from './../../environments/environment';
-import { AuthHttp } from 'angular2-jwt';
+
 import { Injectable } from '@angular/core';
+import { GlobalHttp } from 'app/seguranca/global-http';
 
 @Injectable()
 export class PermissaoService {
@@ -8,15 +9,14 @@ export class PermissaoService {
   permissoesUrl: string;
 
   constructor(
-    private http: AuthHttp
+    private http: GlobalHttp
   ) {
     this.permissoesUrl = `${environment.apiUrl}/permissoes`;
   }
 
   listarTodas(): Promise<any> {
 
-    return this.http.get(this.permissoesUrl)
-      .toPromise()
-      .then(response => response.json());
+    return this.http.get<any>(this.permissoesUrl)
+      .toPromise();
   }
 }
