@@ -1,10 +1,11 @@
+import { GlobalHttp } from './../seguranca/global-http';
 import { AuthService } from './../seguranca/auth.service';
-import { Headers, URLSearchParams } from '@angular/http';
+import { HttpParams, HttpHeaders } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 
 import { environment } from './../../environments/environment';
-import { AuthHttp } from 'angular2-jwt';
-import { Pessoa } from './../core/model';
+
+import { Pessoa, Estado, Cidade } from './../core/model';
 
 import 'rxjs/add/operator/toPromise';
 
@@ -18,36 +19,39 @@ export class PessoaFiltro {
 export class PessoaService {
 
   pessoasUrl: string;
+  cidadesUrl: string;
+  estadosUrl: string;
 
   constructor(
-    private http: AuthHttp,
+    private http: GlobalHttp,
     private auth: AuthService
   ) {
     this.pessoasUrl = `${environment.apiUrl}/pessoas`;
+    this.cidadesUrl = `${environment.apiUrl}/cidades`;
+    this.estadosUrl = `${environment.apiUrl}/estados`;
   }
 
   pesquisar(filtro: PessoaFiltro): Promise<any> {
-    const params = new URLSearchParams();
+    let params = new HttpParams();
 
-    params.set('page', filtro.pagina.toString());
-    params.set('size', filtro.itensPorPagina.toString());
+    params = params.append('page', filtro.pagina.toString());
+    params = params.append('size', filtro.itensPorPagina.toString());
 
     if (filtro.nome) {
-      params.set('nome', filtro.nome);
+      params = params.append('nome', filtro.nome);
     }
 
-    params.set('empresa', this.auth.jwtPayload.empresa.codigo);
+    params = params.append('empresa', this.auth.jwtPayload.empresa.codigo);
 
-    return this.http.get(`${this.pessoasUrl}`,
-        { search: params })
+    return this.http.get<any>(`${this.pessoasUrl}`,
+        { params })
       .toPromise()
       .then(response => {
-        const responseJson = response.json();
-        const pessoas = responseJson.content;
+        const pessoas = response.content;
 
         const resultado = {
           pessoas,
-          total: responseJson.totalElements
+          total: response.totalElements
         };
 
         return resultado;
@@ -55,14 +59,14 @@ export class PessoaService {
   }
 
   listarTodas(): Promise<any> {
-    const params = new URLSearchParams();
+    let params = new HttpParams();
 
-    params.set('empresa', this.auth.jwtPayload.empresa.codigo);
+    params = params.append('empresa', this.auth.jwtPayload.empresa.codigo);
 
-    return this.http.get(this.pessoasUrl,
-      { search: params })
+    return this.http.get<any>(this.pessoasUrl,
+      { params })
       .toPromise()
-      .then(response => response.json().content);
+      .then(response => response.content);
   }
 
   excluir(codigo: number): Promise<void>  {
@@ -73,8 +77,10 @@ export class PessoaService {
   }
 
   mudarStatus(codigo: number, ativo: boolean): Promise<void> {
+    const headers = new HttpHeaders()
+        .append('Content-Type', 'application/json');
 
-    return this.http.put(`${this.pessoasUrl}/${codigo}/ativo`, ativo)
+    return this.http.put(`${this.pessoasUrl}/${codigo}/ativo`, ativo, { headers })
       .toPromise()
       .then(() => null);
   }
@@ -83,19 +89,16 @@ export class PessoaService {
 
     pessoa.empresa.codigo = this.auth.jwtPayload.empresa.codigo;
 
-    return this.http.post(this.pessoasUrl,
-        JSON.stringify(pessoa))
-      .toPromise()
-      .then(response => response.json());
+    return this.http.post<Pessoa>(this.pessoasUrl, pessoa)
+      .toPromise();
   }
 
   atualizar(pessoa: Pessoa): Promise<Pessoa> {
 
-    return this.http.put(`${this.pessoasUrl}/${pessoa.codigo}`,
-        JSON.stringify(pessoa))
+    return this.http.put<Pessoa>(`${this.pessoasUrl}/${pessoa.codigo}`, pessoa)
       .toPromise()
       .then(response => {
-        const pessoaAlterado = response.json() as Pessoa;
+        const pessoaAlterado = response;
 
         return pessoaAlterado;
       });
@@ -103,13 +106,27 @@ export class PessoaService {
 
   buscarPorCodigo(codigo: number): Promise<Pessoa> {
 
-    return this.http.get(`${this.pessoasUrl}/${codigo}`)
+    return this.http.get<Pessoa>(`${this.pessoasUrl}/${codigo}`)
       .toPromise()
       .then(response => {
-        const pessoa = response.json() as Pessoa;
+        const pessoa = response;
 
         return pessoa;
       });
   }
 
+  listarEstados(): Promise<Estado[]> {
+    return this.http.get<Estado[]>(this.estadosUrl)
+      .toPromise();
+  }
+
+  pesquisarCidades(estado): Promise<Cidade[]> {
+    let params = new HttpParams();
+
+    params = params.append('estado', estado);
+
+    return this.http.get<Cidade[]>(this.cidadesUrl,
+      { params })
+      .toPromise();
+  }
 }
