@@ -5,7 +5,7 @@ import { LazyLoadEvent } from 'primeng/components/common/lazyloadevent';
 import { AuthService } from './../../seguranca/auth.service';
 import { ErrorHandlerService } from './../../core/error-handler.service';
 import { ConfirmationService } from 'primeng/components/common/api';
-import { ToastyService } from 'ng2-toasty';
+import { MessageService } from 'primeng/components/common/messageservice';
 import { PessoaFiltro, PessoaService } from './../pessoa.service';
 
 @Component({
@@ -26,7 +26,7 @@ export class PessoasGridComponent {
 
   constructor(
     private pessoaService: PessoaService,
-    private toasty: ToastyService,
+    private messageService: MessageService,
     private confirmation: ConfirmationService,
     private errorHandler: ErrorHandlerService,
     private auth: AuthService) {}
@@ -53,9 +53,10 @@ export class PessoasGridComponent {
           this.consultar.emit();
         } else {
           this.grid.first = 0;
+          this.consultar.emit();
         }
 
-        this.toasty.success('Lançamento excluído com sucesso!');
+        this.messageService.add({ severity: 'success', detail: 'Lançamento excluído com sucesso!'});
       })
       .catch(erro => this.errorHandler.handle(erro));
   }
@@ -69,7 +70,8 @@ export class PessoasGridComponent {
         const acao = novoStatus ? 'ativada' : 'desativada';
 
         pessoa.ativo = novoStatus;
-        this.toasty.success(`Pessoa ${acao} com sucesso!`);
+
+        this.messageService.add({ severity: 'success', detail: `Pessoa ${acao} com sucesso!`});
       })
       .catch(erro => this.errorHandler.handle(erro));
   }
